Got the classify option working
This required adding a load() function to the ItchJam class, which made me realize I needed to be able to init an empty ItchJam... it's a long story. There's a good chance that at the end of all this I'll stop pretending it's a pure dataclass.

Still, coming along nicely. Most of the core functionality is there; I just have output left to do.

Also I am going to move the crawling routines into the ItchJamList class in due time.