First steps, refactoring DB
OK, so we need a list of jam owners and owner IDs for each jam -- https://physicalgamejams.carrd.co has that, so we should. dataset is not going to handle that for us elegantly (no, I'm not gonna save tuples as strings and then unwind them). dataset can't handle JSON types for SQLite.

Fine fine fine, we'll drop dataset and go back to pure old SQLite and learn how JSON works, this'll be fun. It actually is kind of fun but it's a bit of work.

So far I've got a save routine that works and I've modified auto_classify to work. But I really need to swap over to treating owner info as a tuple. This is taking long enough so that I've grown up and am using a new branch, exciting.