Removing dataclass
It's a shame but I wasn't using it right anyhow and I am too weary of this code to figure out how to type something as a list. I can be a better programmer in this aspect on the next project.

The code isn't currently working but I gotta go have dinner, let's save where we are.