Added sortable tables
Used https://github.com/tofsjonas/sortable -- easy and quick. Added an
attribute to each date cell with a more sortable date. Sorts on end date
when the page is loaded -- this is redundant since I generate the table
sorted that way to start with, but it's a quick and dirty way to get the
visual indicator in place.