Finished adding old jams to the site
Finished up with the fiddly bits, tested, looks good. I need to make
the number of items per page configurable, but that's a job for another
time. It's probably also time to think about splitting up the list
function, too...

Thanks to a gentleman named Rob Smith for prodding me on this.